Split into some crates
mchprs_blocks: block, item, and block entity structures. (the actual `Block` enum will remain in mchprs_core for now until the functionality is properly split from the data)
mchprs_save_data: plot data structure and old data conversion
mchprs_world: this will eventually have shared world functionality
mchprs_utils: some utility macros
mchprs_core: the original crate and where most of the functionality remains

This is a step forward in improving mchprs's code quality (and compile times).
