+++ crates/blocks/src/block_entities.rs
@@ -1,0 +1,232 @@
+use mchprs_utils::{map, nbt_unwrap_val};
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+use std::str::FromStr;
+
+use crate::items::Item;
+
+/// A single item in an inventory
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct InventoryEntry {
+    pub id: u32,
+    pub slot: i8,
+    pub count: i8,
+    pub nbt: Option<Vec<u8>>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct SignBlockEntity {
+    pub rows: [String; 4],
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+pub enum ContainerType {
+    Furnace,
+    Barrel,
+    Hopper,
+}
+
+impl FromStr for ContainerType {
+    type Err = ();
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Ok(match s {
+            "barrel" => ContainerType::Barrel,
+            "furnace" => ContainerType::Furnace,
+            "hopper" => ContainerType::Hopper,
+            _ => return Err(()),
+        })
+    }
+}
+
+impl ToString for ContainerType {
+    fn to_string(&self) -> String {
+        match self {
+            ContainerType::Furnace => "minecraft:furnace",
+            ContainerType::Barrel => "minecraft:barrel",
+            ContainerType::Hopper => "minecraft:hopper",
+        }
+        .to_owned()
+    }
+}
+
+impl ContainerType {
+    pub fn num_slots(self) -> u8 {
+        match self {
+            ContainerType::Furnace => 3,
+            ContainerType::Barrel => 27,
+            ContainerType::Hopper => 5,
+        }
+    }
+
+    pub fn window_type(self) -> u8 {
+        // https://wiki.vg/Inventory
+        match self {
+            ContainerType::Furnace => 13,
+            ContainerType::Barrel => 2,
+            ContainerType::Hopper => 15,
+        }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub enum BlockEntity {
+    Comparator {
+        output_strength: u8,
+    },
+    Container {
+        comparator_override: u8,
+        inventory: Vec<InventoryEntry>,
+        ty: ContainerType,
+    },
+    Sign(Box<SignBlockEntity>),
+}
+
+impl BlockEntity {
+    /// The protocol id for the block entity
+    pub fn ty(&self) -> i32 {
+        match self {
+            BlockEntity::Comparator { .. } => 17,
+            BlockEntity::Container { ty, .. } => match ty {
+                ContainerType::Furnace => 0,
+                ContainerType::Barrel => 25,
+                ContainerType::Hopper => 16,
+            },
+            BlockEntity::Sign(_) => 7,
+        }
+    }
+
+    fn load_container(slots_nbt: &[nbt::Value], ty: ContainerType) -> Option<BlockEntity> {
+        use nbt::Value;
+        let num_slots = ty.num_slots();
+        let mut fullness_sum: f32 = 0.0;
+        let mut inventory = Vec::new();
+        for item in slots_nbt {
+            let item_compound = nbt_unwrap_val!(item, Value::Compound);
+            let count = nbt_unwrap_val!(item_compound["Count"], Value::Byte);
+            let slot = nbt_unwrap_val!(item_compound["Slot"], Value::Byte);
+            let namespaced_name = nbt_unwrap_val!(
+                item_compound
+                    .get("Id")
+                    .or_else(|| item_compound.get("id"))?,
+                Value::String
+            );
+            let item_type = Item::from_name(namespaced_name.split(':').last()?);
+
+            let mut blob = nbt::Blob::new();
+            for (k, v) in item_compound {
+                blob.insert(k, v.clone()).unwrap();
+            }
+            let mut data = Vec::new();
+            blob.to_writer(&mut data).unwrap();
+
+            let tag = match item_compound.get("tag") {
+                Some(nbt::Value::Compound(map)) => {
+                    let mut blob = nbt::Blob::new();
+                    for (k, v) in map {
+                        blob.insert(k, v.clone()).unwrap();
+                    }
+
+                    let mut data = Vec::new();
+                    blob.to_writer(&mut data).unwrap();
+                    Some(data)
+                }
+                _ => None,
+            };
+            inventory.push(InventoryEntry {
+                slot,
+                count,
+                id: item_type.unwrap_or(Item::Redstone {}).get_id(),
+                nbt: tag,
+            });
+
+            fullness_sum += count as f32 / item_type.map_or(64, Item::max_stack_size) as f32;
+        }
+        Some(BlockEntity::Container {
+            comparator_override: (if fullness_sum > 0.0 { 1.0 } else { 0.0 }
+                + (fullness_sum / num_slots as f32) * 14.0)
+                .floor() as u8,
+            inventory,
+            ty,
+        })
+    }
+
+    pub fn from_nbt(nbt: &HashMap<String, nbt::Value>) -> Option<BlockEntity> {
+        use nbt::Value;
+        let id = nbt_unwrap_val!(&nbt.get("Id").or_else(|| nbt.get("id"))?, Value::String);
+        match id.as_ref() {
+            "minecraft:comparator" => Some(BlockEntity::Comparator {
+                output_strength: *nbt_unwrap_val!(&nbt["OutputSignal"], Value::Int) as u8,
+            }),
+            "minecraft:furnace" => BlockEntity::load_container(
+                nbt_unwrap_val!(&nbt["Items"], Value::List),
+                ContainerType::Furnace,
+            ),
+            "minecraft:barrel" => BlockEntity::load_container(
+                nbt_unwrap_val!(&nbt["Items"], Value::List),
+                ContainerType::Barrel,
+            ),
+            "minecraft:hopper" => BlockEntity::load_container(
+                nbt_unwrap_val!(&nbt["Items"], Value::List),
+                ContainerType::Hopper,
+            ),
+            "minecraft:sign" => Some({
+                BlockEntity::Sign(Box::new(SignBlockEntity {
+                    rows: [
+                        // This cloning is really dumb
+                        nbt_unwrap_val!(nbt["Text1"].clone(), Value::String),
+                        nbt_unwrap_val!(nbt["Text2"].clone(), Value::String),
+                        nbt_unwrap_val!(nbt["Text3"].clone(), Value::String),
+                        nbt_unwrap_val!(nbt["Text4"].clone(), Value::String),
+                    ],
+                }))
+            }),
+            _ => None,
+        }
+    }
+
+    pub fn to_nbt(&self, sign_only: bool) -> Option<nbt::Blob> {
+        if sign_only && !matches!(self, BlockEntity::Sign(_)) {
+            return None;
+        }
+
+        use nbt::Value;
+        match self {
+            BlockEntity::Sign(sign) => Some({
+                let mut blob = nbt::Blob::new();
+                let [r1, r2, r3, r4] = sign.rows.clone();
+                let _ = blob.insert("Text1", Value::String(r1));
+                let _ = blob.insert("Text2", Value::String(r2));
+                let _ = blob.insert("Text3", Value::String(r3));
+                let _ = blob.insert("Text4", Value::String(r4));
+                let _ = blob.insert("id", Value::String("minecraft:sign".to_owned()));
+                blob
+            }),
+            BlockEntity::Comparator { output_strength } => Some({
+                let mut blob = nbt::Blob::new();
+                let _ = blob.insert("OutputSignal", Value::Int(*output_strength as i32));
+                let _ = blob.insert("id", Value::String("minecraft:comparator".to_owned()));
+                blob
+            }),
+            BlockEntity::Container { inventory, ty, .. } => Some({
+                let mut blob = nbt::Blob::new();
+                let _ = blob.insert("id", Value::String(ty.to_string()));
+                let mut items = Vec::new();
+                for entry in inventory {
+                    let nbt = map! {
+                        "Count".to_owned() => nbt::Value::Byte(entry.count),
+                        "id".to_owned() => nbt::Value::String("minecraft:".to_string() + Item::from_id(entry.id).get_name()),
+                        "Slot".to_owned() => nbt::Value::Byte(entry.slot)
+                    };
+                    // TODO: item nbt data in containers
+                    // if let Some(tag) = &entry.nbt {
+                    //     let blob = nbt::Blob::from_reader(&mut Cursor::new(tag)).unwrap();
+                    // }
+                    items.push(nbt::Value::Compound(nbt));
+                }
+                let _ = blob.insert("Items", Value::List(items));
+                blob
+            }),
+        }
+    }
+}
