+++ crates/core/Cargo.toml
@@ -21,14 +21,12 @@ mchprs_proc_macros = { path = "../proc_macros" }
 toml = "0.5"
 flate2 = "1"
 byteorder = "1.4"
-# hematite-nbt = "0.5"
-hematite-nbt = { git = "https://github.com/PistonDevelopers/hematite_nbt" }
+hematite-nbt = "0.5"
 bitflags = "1.2"
 serde = "1"
 serde_json = "1.0"
 md5 = "0.7"
 bus = "2.2"
-bincode = "1.3"
 ctrlc = { version = "3.1", features = ["termination"] }
 log = "0.4"
 fern = { version = "0.6", features = ["colored"] }
@@ -48,7 +46,12 @@ tokio = { version = "1", features = ["rt-multi-thread"] }
 reqwest = { version = "0.11", features = ["json"] }
 itertools = "0.10"
 impls = "1"
+bincode = "1.3"
 redpiler_graph = { path = "../redpiler_graph" }
+mchprs_save_data = { path = "../save_data" }
+mchprs_blocks = { path = "../blocks" }
+mchprs_world = { path = "../world" }
+mchprs_utils = { path = "../utils" }
 
 [features]
 jit_cranelift = ["cranelift", "cranelift-jit", "cranelift-module"]
