+++ crates/core/src/blocks/mod.rs
@@ -1,11 +1,16 @@
 mod redstone;
 
-use crate::items::{ActionResult, InventoryEntry, Item, UseOnBlockContext};
+use crate::items::{ActionResult, UseOnBlockContext};
 use crate::player::Player;
-use crate::world::{TickPriority, World};
-use mchprs_proc_macros::{BlockProperty, BlockTransform};
+use crate::world::World;
+use mchprs_blocks::block_entities::BlockEntity;
+use mchprs_blocks::items::Item;
+use mchprs_blocks::{
+    BlockColorVariant, BlockDirection, BlockFace, BlockFacing, BlockPos, BlockProperty,
+};
+use mchprs_proc_macros::BlockTransform;
+use mchprs_world::TickPriority;
 pub use redstone::*;
-use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::str::FromStr;
 
@@ -47,410 +52,6 @@ impl<T> BlockTransform for T {
     default fn flip(&mut self, _dir: crate::blocks::FlipDirection) {}
 }
 
-trait BlockProperty: Sized {
-    fn encode(self, props: &mut HashMap<&'static str, String>, name: &'static str);
-    fn decode(&mut self, props: &HashMap<&str, &str>, name: &str);
-}
-
-impl<T> BlockProperty for T
-where
-    T: ToString + FromStr,
-{
-    fn encode(self, props: &mut HashMap<&'static str, String>, name: &'static str) {
-        props.insert(name, self.to_string());
-    }
-
-    fn decode(&mut self, props: &HashMap<&str, &str>, name: &str) {
-        if let Some(&str) = props.get(name) {
-            if let Ok(val) = str.parse() {
-                *self = val;
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SignBlockEntity {
-    pub rows: [String; 4],
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-pub enum ContainerType {
-    Furnace,
-    Barrel,
-    Hopper,
-}
-
-impl FromStr for ContainerType {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(match s {
-            "barrel" => ContainerType::Barrel,
-            "furnace" => ContainerType::Furnace,
-            "hopper" => ContainerType::Hopper,
-            _ => return Err(()),
-        })
-    }
-}
-
-impl ToString for ContainerType {
-    fn to_string(&self) -> String {
-        match self {
-            ContainerType::Furnace => "minecraft:furnace",
-            ContainerType::Barrel => "minecraft:barrel",
-            ContainerType::Hopper => "minecraft:hopper",
-        }
-        .to_owned()
-    }
-}
-
-impl ContainerType {
-    pub fn num_slots(self) -> u8 {
-        match self {
-            ContainerType::Furnace => 3,
-            ContainerType::Barrel => 27,
-            ContainerType::Hopper => 5,
-        }
-    }
-
-    pub fn window_type(self) -> u8 {
-        // https://wiki.vg/Inventory
-        match self {
-            ContainerType::Furnace => 13,
-            ContainerType::Barrel => 2,
-            ContainerType::Hopper => 15,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum BlockEntity {
-    Comparator {
-        output_strength: u8,
-    },
-    Container {
-        comparator_override: u8,
-        inventory: Vec<InventoryEntry>,
-        ty: ContainerType,
-    },
-    Sign(Box<SignBlockEntity>),
-}
-
-impl BlockEntity {
-    /// The protocol id for the block entity
-    pub fn ty(&self) -> i32 {
-        match self {
-            BlockEntity::Comparator { .. } => 17,
-            BlockEntity::Container { ty, .. } => match ty {
-                ContainerType::Furnace => 0,
-                ContainerType::Barrel => 25,
-                ContainerType::Hopper => 16,
-            },
-            BlockEntity::Sign(_) => 7,
-        }
-    }
-
-    fn load_container(slots_nbt: &[nbt::Value], ty: ContainerType) -> Option<BlockEntity> {
-        use nbt::Value;
-        let num_slots = ty.num_slots();
-        let mut fullness_sum: f32 = 0.0;
-        let mut inventory = Vec::new();
-        for item in slots_nbt {
-            let item_compound = nbt_unwrap_val!(item, Value::Compound);
-            let count = nbt_unwrap_val!(item_compound["Count"], Value::Byte);
-            let slot = nbt_unwrap_val!(item_compound["Slot"], Value::Byte);
-            let namespaced_name = nbt_unwrap_val!(
-                item_compound
-                    .get("Id")
-                    .or_else(|| item_compound.get("id"))?,
-                Value::String
-            );
-            let item_type = Item::from_name(namespaced_name.split(':').last()?);
-
-            let mut blob = nbt::Blob::new();
-            for (k, v) in item_compound {
-                blob.insert(k, v.clone()).unwrap();
-            }
-            let mut data = Vec::new();
-            blob.to_writer(&mut data).unwrap();
-
-            let tag = match item_compound.get("tag") {
-                Some(nbt::Value::Compound(map)) => {
-                    let mut blob = nbt::Blob::new();
-                    for (k, v) in map {
-                        blob.insert(k, v.clone()).unwrap();
-                    }
-
-                    let mut data = Vec::new();
-                    blob.to_writer(&mut data).unwrap();
-                    Some(data)
-                }
-                _ => None,
-            };
-            inventory.push(InventoryEntry {
-                slot,
-                count,
-                id: item_type.unwrap_or(Item::Redstone {}).get_id(),
-                nbt: tag,
-            });
-
-            fullness_sum += count as f32 / item_type.map_or(64, Item::max_stack_size) as f32;
-        }
-        Some(BlockEntity::Container {
-            comparator_override: (if fullness_sum > 0.0 { 1.0 } else { 0.0 }
-                + (fullness_sum / num_slots as f32) * 14.0)
-                .floor() as u8,
-            inventory,
-            ty,
-        })
-    }
-
-    pub fn from_nbt(nbt: &HashMap<String, nbt::Value>) -> Option<BlockEntity> {
-        use nbt::Value;
-        let id = nbt_unwrap_val!(&nbt.get("Id").or_else(|| nbt.get("id"))?, Value::String);
-        match id.as_ref() {
-            "minecraft:comparator" => Some(BlockEntity::Comparator {
-                output_strength: *nbt_unwrap_val!(&nbt["OutputSignal"], Value::Int) as u8,
-            }),
-            "minecraft:furnace" => BlockEntity::load_container(
-                nbt_unwrap_val!(&nbt["Items"], Value::List),
-                ContainerType::Furnace,
-            ),
-            "minecraft:barrel" => BlockEntity::load_container(
-                nbt_unwrap_val!(&nbt["Items"], Value::List),
-                ContainerType::Barrel,
-            ),
-            "minecraft:hopper" => BlockEntity::load_container(
-                nbt_unwrap_val!(&nbt["Items"], Value::List),
-                ContainerType::Hopper,
-            ),
-            "minecraft:sign" => Some({
-                BlockEntity::Sign(Box::new(SignBlockEntity {
-                    rows: [
-                        // This cloning is really dumb
-                        nbt_unwrap_val!(nbt["Text1"].clone(), Value::String),
-                        nbt_unwrap_val!(nbt["Text2"].clone(), Value::String),
-                        nbt_unwrap_val!(nbt["Text3"].clone(), Value::String),
-                        nbt_unwrap_val!(nbt["Text4"].clone(), Value::String),
-                    ],
-                }))
-            }),
-            _ => None,
-        }
-    }
-
-    pub fn to_nbt(&self, sign_only: bool) -> Option<nbt::Blob> {
-        if sign_only && !matches!(self, BlockEntity::Sign(_)) {
-            return None;
-        }
-
-        use nbt::Value;
-        match self {
-            BlockEntity::Sign(sign) => Some({
-                let mut blob = nbt::Blob::new();
-                let [r1, r2, r3, r4] = sign.rows.clone();
-                let _ = blob.insert("Text1", Value::String(r1));
-                let _ = blob.insert("Text2", Value::String(r2));
-                let _ = blob.insert("Text3", Value::String(r3));
-                let _ = blob.insert("Text4", Value::String(r4));
-                let _ = blob.insert("id", Value::String("minecraft:sign".to_owned()));
-                blob
-            }),
-            BlockEntity::Comparator { output_strength } => Some({
-                let mut blob = nbt::Blob::new();
-                let _ = blob.insert("OutputSignal", Value::Int(*output_strength as i32));
-                let _ = blob.insert("id", Value::String("minecraft:comparator".to_owned()));
-                blob
-            }),
-            BlockEntity::Container { inventory, ty, .. } => Some({
-                let mut blob = nbt::Blob::new();
-                let _ = blob.insert("id", Value::String(ty.to_string()));
-                let mut items = Vec::new();
-                for entry in inventory {
-                    let nbt = map! {
-                        "Count".to_owned() => nbt::Value::Byte(entry.count),
-                        "id".to_owned() => nbt::Value::String("minecraft:".to_string() + Item::from_id(entry.id).get_name()),
-                        "Slot".to_owned() => nbt::Value::Byte(entry.slot)
-                    };
-                    // TODO: item nbt data in containers
-                    // if let Some(tag) = &entry.nbt {
-                    //     let blob = nbt::Blob::from_reader(&mut Cursor::new(tag)).unwrap();
-                    // }
-                    items.push(nbt::Value::Compound(nbt));
-                }
-                let _ = blob.insert("Items", Value::List(items));
-                blob
-            }),
-        }
-    }
-}
-
-#[derive(PartialEq, Eq, Copy, Clone, Debug, Serialize, Deserialize, Hash)]
-pub struct BlockPos {
-    pub x: i32,
-    pub y: i32,
-    pub z: i32,
-}
-
-impl BlockPos {
-    pub const fn new(x: i32, y: i32, z: i32) -> BlockPos {
-        BlockPos { x, y, z }
-    }
-
-    pub fn zero() -> BlockPos {
-        BlockPos::new(0, 0, 0)
-    }
-
-    pub fn offset(self, face: BlockFace) -> BlockPos {
-        match face {
-            BlockFace::Bottom => BlockPos::new(self.x, self.y.saturating_sub(1), self.z),
-            BlockFace::Top => BlockPos::new(self.x, self.y + 1, self.z),
-            BlockFace::North => BlockPos::new(self.x, self.y, self.z - 1),
-            BlockFace::South => BlockPos::new(self.x, self.y, self.z + 1),
-            BlockFace::West => BlockPos::new(self.x - 1, self.y, self.z),
-            BlockFace::East => BlockPos::new(self.x + 1, self.y, self.z),
-        }
-    }
-
-    pub fn max(self, other: BlockPos) -> BlockPos {
-        BlockPos {
-            x: std::cmp::max(self.x, other.x),
-            y: std::cmp::max(self.y, other.y),
-            z: std::cmp::max(self.z, other.z),
-        }
-    }
-
-    pub fn min(self, other: BlockPos) -> BlockPos {
-        BlockPos {
-            x: std::cmp::min(self.x, other.x),
-            y: std::cmp::min(self.y, other.y),
-            z: std::cmp::min(self.z, other.z),
-        }
-    }
-}
-
-impl std::ops::Sub for BlockPos {
-    type Output = BlockPos;
-
-    fn sub(self, rhs: BlockPos) -> BlockPos {
-        BlockPos {
-            x: self.x - rhs.x,
-            y: self.y - rhs.y,
-            z: self.z - rhs.z,
-        }
-    }
-}
-
-impl std::ops::Add for BlockPos {
-    type Output = BlockPos;
-
-    fn add(self, rhs: BlockPos) -> BlockPos {
-        BlockPos {
-            x: self.x + rhs.x,
-            y: self.y + rhs.y,
-            z: self.z + rhs.z,
-        }
-    }
-}
-
-impl std::ops::Mul<i32> for BlockPos {
-    type Output = BlockPos;
-
-    fn mul(self, rhs: i32) -> BlockPos {
-        BlockPos {
-            x: self.x * rhs,
-            y: self.y * rhs,
-            z: self.z * rhs,
-        }
-    }
-}
-
-impl std::fmt::Display for BlockPos {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "({}, {}, {})", self.x, self.y, self.z)
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum BlockDirection {
-    North,
-    South,
-    East,
-    West,
-}
-
-impl BlockDirection {
-    pub fn opposite(self) -> BlockDirection {
-        use BlockDirection::*;
-        match self {
-            North => South,
-            South => North,
-            East => West,
-            West => East,
-        }
-    }
-
-    pub fn block_face(self) -> BlockFace {
-        use BlockDirection::*;
-        match self {
-            North => BlockFace::North,
-            South => BlockFace::South,
-            East => BlockFace::East,
-            West => BlockFace::West,
-        }
-    }
-
-    pub fn block_facing(self) -> BlockFacing {
-        use BlockDirection::*;
-        match self {
-            North => BlockFacing::North,
-            South => BlockFacing::South,
-            East => BlockFacing::East,
-            West => BlockFacing::West,
-        }
-    }
-
-    pub fn from_id(id: u32) -> BlockDirection {
-        match id {
-            0 => BlockDirection::North,
-            1 => BlockDirection::South,
-            2 => BlockDirection::West,
-            3 => BlockDirection::East,
-            _ => panic!("Invalid BlockDirection"),
-        }
-    }
-
-    fn get_id(self) -> u32 {
-        match self {
-            BlockDirection::North => 0,
-            BlockDirection::South => 1,
-            BlockDirection::West => 2,
-            BlockDirection::East => 3,
-        }
-    }
-
-    pub fn rotate(self) -> BlockDirection {
-        use BlockDirection::*;
-        match self {
-            North => East,
-            East => South,
-            South => West,
-            West => North,
-        }
-    }
-
-    pub fn rotate_ccw(self) -> BlockDirection {
-        use BlockDirection::*;
-        match self {
-            North => West,
-            West => South,
-            South => East,
-            East => North,
-        }
-    }
-}
-
 impl BlockTransform for BlockDirection {
     fn flip(&mut self, dir: FlipDirection) {
         match dir {
@@ -475,242 +76,6 @@ impl BlockTransform for BlockDirection {
             BlockDirection::West => BlockDirection::North,
         }
     }
-}
-
-impl FromStr for BlockDirection {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(match s {
-            "north" => BlockDirection::North,
-            "south" => BlockDirection::South,
-            "east" => BlockDirection::East,
-            "west" => BlockDirection::West,
-            _ => return Err(()),
-        })
-    }
-}
-
-impl ToString for BlockDirection {
-    fn to_string(&self) -> String {
-        match self {
-            BlockDirection::North => "north".to_owned(),
-            BlockDirection::South => "south".to_owned(),
-            BlockDirection::East => "east".to_owned(),
-            BlockDirection::West => "west".to_owned(),
-        }
-    }
-}
-
-impl Default for BlockDirection {
-    fn default() -> Self {
-        BlockDirection::West
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum BlockFace {
-    Bottom,
-    Top,
-    North,
-    South,
-    West,
-    East,
-}
-
-impl BlockFace {
-    pub fn from_id(id: u32) -> BlockFace {
-        match id {
-            0 => BlockFace::Bottom,
-            1 => BlockFace::Top,
-            2 => BlockFace::North,
-            3 => BlockFace::South,
-            4 => BlockFace::West,
-            5 => BlockFace::East,
-            _ => BlockFace::West,
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum BlockFacing {
-    North,
-    East,
-    South,
-    West,
-    Up,
-    Down,
-}
-
-impl BlockFacing {
-    fn from_id(id: u32) -> BlockFacing {
-        match id {
-            0 => BlockFacing::North,
-            1 => BlockFacing::East,
-            2 => BlockFacing::South,
-            3 => BlockFacing::West,
-            4 => BlockFacing::Up,
-            5 => BlockFacing::Down,
-            _ => BlockFacing::West,
-        }
-    }
-
-    fn get_id(self) -> u32 {
-        match self {
-            BlockFacing::North => 0,
-            BlockFacing::East => 1,
-            BlockFacing::South => 2,
-            BlockFacing::West => 3,
-            BlockFacing::Up => 4,
-            BlockFacing::Down => 5,
-        }
-    }
-
-    pub fn offset_pos(self, mut pos: BlockPos, n: i32) -> BlockPos {
-        match self {
-            BlockFacing::North => pos.z -= n,
-            BlockFacing::South => pos.z += n,
-            BlockFacing::East => pos.x += n,
-            BlockFacing::West => pos.x -= n,
-            BlockFacing::Up => pos.y += n,
-            BlockFacing::Down => pos.y -= n,
-        }
-        pos
-    }
-
-    pub fn rotate(self) -> BlockFacing {
-        use BlockFacing::*;
-        match self {
-            North => East,
-            East => South,
-            South => West,
-            West => North,
-            other => other,
-        }
-    }
-
-    pub fn rotate_ccw(self) -> BlockFacing {
-        use BlockFacing::*;
-        match self {
-            North => West,
-            West => South,
-            South => East,
-            East => North,
-            other => other,
-        }
-    }
-}
-
-impl ToString for BlockFacing {
-    fn to_string(&self) -> String {
-        match self {
-            BlockFacing::North => "north".to_owned(),
-            BlockFacing::South => "south".to_owned(),
-            BlockFacing::East => "east".to_owned(),
-            BlockFacing::West => "west".to_owned(),
-            BlockFacing::Up => "up".to_owned(),
-            BlockFacing::Down => "down".to_owned(),
-        }
-    }
-}
-
-impl FromStr for BlockFacing {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(match s {
-            "north" => BlockFacing::North,
-            "south" => BlockFacing::South,
-            "east" => BlockFacing::East,
-            "west" => BlockFacing::West,
-            "up" => BlockFacing::Up,
-            "down" => BlockFacing::Down,
-            _ => return Err(()),
-        })
-    }
-}
-
-impl Default for BlockFacing {
-    fn default() -> Self {
-        BlockFacing::West
-    }
-}
-
-impl BlockFace {
-    pub fn values() -> [BlockFace; 6] {
-        use BlockFace::*;
-        [Top, Bottom, North, South, East, West]
-    }
-
-    pub fn is_horizontal(self) -> bool {
-        use BlockFace::*;
-        matches!(self, North | South | East | West)
-    }
-
-    pub fn to_direction(self) -> BlockDirection {
-        match self {
-            BlockFace::North => BlockDirection::North,
-            BlockFace::South => BlockDirection::South,
-            BlockFace::East => BlockDirection::East,
-            BlockFace::West => BlockDirection::West,
-            _ => BlockDirection::West,
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum BlockColorVariant {
-    White = 0,
-    Orange = 1,
-    Magenta = 2,
-    LightBlue = 3,
-    Yellow = 4,
-    Lime = 5,
-    Pink = 6,
-    Gray = 7,
-    LightGray = 8,
-    Cyan = 9,
-    Purple = 10,
-    Blue = 11,
-    Brown = 12,
-    Green = 13,
-    Red = 14,
-    Black = 15,
-}
-
-impl BlockColorVariant {
-    pub fn get_id(self) -> u32 {
-        self as u32
-    }
-
-    pub fn from_id(id: u32) -> BlockColorVariant {
-        use BlockColorVariant::*;
-        match id {
-            0 => White,
-            1 => Orange,
-            2 => Magenta,
-            3 => LightBlue,
-            4 => Yellow,
-            5 => Lime,
-            6 => Pink,
-            7 => Gray,
-            8 => LightGray,
-            9 => Cyan,
-            10 => Purple,
-            11 => Blue,
-            12 => Brown,
-            13 => Green,
-            14 => Red,
-            15 => Black,
-            _ => unreachable!(),
-        }
-    }
-}
-
-impl BlockProperty for BlockColorVariant {
-    // Don't encode: the color is encoded in the block name
-    fn encode(self, _props: &mut HashMap<&'static str, String>, _name: &'static str) {}
-    fn decode(&mut self, _props: &HashMap<&str, &str>, _name: &str) {}
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
