+++ crates/core/src/plot/worldedit/execute.rs
@@ -1,13 +1,15 @@
 use super::*;
-use crate::blocks::{Block, BlockFace, BlockFacing, BlockPos, FlipDirection, RotateAmt};
+use crate::blocks::{Block, FlipDirection, RotateAmt};
 use crate::chat::{ChatComponentBuilder, ColorCode};
 use crate::config::CONFIG;
-use crate::items::{InventoryEntry, Item, ItemStack};
 use crate::network::packets::clientbound::*;
 use crate::network::packets::SlotData;
 use crate::player::PacketSender;
 use crate::utils::HyphenatedUUID;
 use log::error;
+use mchprs_blocks::block_entities::InventoryEntry;
+use mchprs_blocks::items::{Item, ItemStack};
+use mchprs_blocks::{BlockFace, BlockFacing, BlockPos};
 use schematic::{load_schematic, save_schematic};
 use std::time::Instant;
