+++ crates/core/src/plot/worldedit/mod.rs
@@ -4,11 +4,14 @@ mod execute;
 mod schematic;
 
 use super::{Plot, PlotWorld};
-use crate::blocks::{Block, BlockEntity, BlockFacing, BlockPos, ContainerType};
+use crate::blocks::Block;
 use crate::player::{PacketSender, Player, PlayerPos};
 use crate::world::storage::PalettedBitBuffer;
 use crate::world::World;
 use execute::*;
+use mchprs_blocks::block_entities::{BlockEntity, ContainerType};
+use mchprs_blocks::{BlockFacing, BlockPos};
+use mchprs_utils::map;
 use rand::Rng;
 use regex::Regex;
 use std::collections::HashMap;
