+++ crates/core/src/utils.rs
@@ -33,30 +33,6 @@ impl<T: Serialize> NBTMap<T> {
     }
 }
 
-/// An easy way to create HashMaps
-macro_rules! map(
-    { $($key:expr => $value:expr),+ } => {
-        {
-            let mut m = ::std::collections::HashMap::new();
-            $(
-                m.insert($key, $value);
-            )+
-            m
-        }
-     };
-);
-
-macro_rules! nbt_unwrap_val {
-    // I'm not sure if path is the right type here.
-    // It works though!
-    ($e:expr, $p:path) => {
-        match $e {
-            $p(val) => val,
-            _ => return None,
-        }
-    };
-}
-
 #[derive(Debug)]
 pub struct HyphenatedUUID(pub u128);
