+++ crates/save_data/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "mchprs_save_data"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+byteorder = "1.4"
+bincode = "1.3"
+serde = "1"
+thiserror = "1"
+mchprs_world = { path = "../world" }
+mchprs_blocks = { path = "../blocks" }
