+++ crates/save_data/src/plot_data/fixer/pre_header.rs
@@ -1,0 +1,7 @@
+//! If the mchprs plot header is missing, it's possible it can still be a
+//! plot save file. For mchprs versions targetting 1.17.1 and below, we did
+//! not have a file header.
+
+pub fn try_fix(data: &[u8]) {
+
+}
