+++ crates/utils/src/lib.rs
@@ -1,0 +1,25 @@
+/// An easy way to create HashMaps
+#[macro_export]
+macro_rules! map(
+    { $($key:expr => $value:expr),+ } => {
+        {
+            let mut m = ::std::collections::HashMap::new();
+            $(
+                m.insert($key, $value);
+            )+
+            m
+        }
+     };
+);
+
+#[macro_export]
+macro_rules! nbt_unwrap_val {
+    // I'm not sure if path is the right type here.
+    // It works though!
+    ($e:expr, $p:path) => {
+        match $e {
+            $p(val) => val,
+            _ => return None,
+        }
+    };
+}
